idk about dlg solve
