+++ Misc/data-reconstruction-federated-learning/writeup/WRITEUP.md
@@ -5,8 +5,6 @@
 
 ## Writeup
 
-[Exploit](../src/reconstruct.py) based off this [paper](https://proceedings.neurips.cc/paper/2019/hash/60a6c4002cc7b29142def8871531281a-Abstract.html)
+The [exploit](../src/solve.py) based off this [paper](https://proceedings.neurips.cc/paper/2019/hash/60a6c4002cc7b29142def8871531281a-Abstract.html)
 
-Flag:
-
-`UiTHack23{t1m3_t0_l34rn}`
+[Flag](../src/flag.txt)
